refactor BatchStatusTest to use parameterized tests
